requirements.txt
psycopg2
netifaces
django
python-dotenv
django-rest-framework